+++ src/components/collage/CollageNetwork.vue
